Changed "Where to?" preference to live inside the Select Preferences option on the home screen.
